+++ tests/init/InstallationTest.php
@@ -3,7 +3,7 @@
 namespace lucatume\WPBrowser\Tests;
 
 use lucatume\WPBrowser\Template\Wpbrowser;
-use lucatume\WPBrowser\Tests\Traits\WithUopz;
+use lucatume\WPBrowser\Tests\Traits\UopzFunctions;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Output\NullOutput;
 use tad\Codeception\SnapshotAssertions\SnapshotAssertions;
@@ -13,7 +13,7 @@ require_once __DIR__ . '/BaseTest.php';
 class InstallationTest extends BaseTest
 {
     use SnapshotAssertions;
-    use WithUopz;
+    use UopzFunctions;
 
     /**
      * It should correctly scaffold quiet installation